Move filter code to accessible utilities location
